requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
black==24.2.0
click==8.1.7
django==4.2.10
djangorestframework==3.14.0
exceptiongroup==1.2.0
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
pip==24.0
platformdirs==4.2.0
pluggy==1.4.0
ply==3.11
pydantic==2.6.1
pydantic-core==2.16.2
pyright==1.1.351
pytest==8.0.1
python-dateutil==2.8.2
pytz==2024.1
rule-engine==4.3.1
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.9.0
wheel==0.42.0
